refactor: 로컬스토리지 작업 컴포넌트화
