+++ src/app/(route)/(community)/news/news-detail/[id]/page.tsx
@@ -14,6 +14,7 @@ import { NewsItemType } from "@/app/_constants/newsItemType";
 import useGetNewsInfoData from "@/_hooks/useGetNewsInfoData";
 import PostNavigation from "../../../_components/PostNavigation";
 import useSortedNewsDataList from "@/_hooks/useSortedPosts";
+import { useReadNews } from "@/_hooks/useReadNews";
 
 const Page = ({ params }: { params: Promise<{ id: string }> }) => {
   const [orderType, setOrderType] = useState<"DATE" | "COMMENT" | "VIEW">(
@@ -26,6 +27,7 @@ const Page = ({ params }: { params: Promise<{ id: string }> }) => {
   const { data: newsListData } = useSortedNewsDataList({ orderType, pageNum });
   const sliceNewsListData = newsListData ? newsListData.slice(0, 3) : [];
   const updatedImgUrl = data?.thumbImg?.replace("type=w140", "type=w360"); // 뉴스 상세페이지 들어갔을때 이미지 화질 올리는 로직
+  const { isRead, handleRead } = useReadNews(sliceNewsListData?.id);
 
   const changeCategory = (category: string) => {
     switch (category) {
@@ -167,7 +169,7 @@ const Page = ({ params }: { params: Promise<{ id: string }> }) => {
       </div>
       <div className="w-[720px] min-h-[348px] rounded-b-[5px] overflow-hidden shadow-md">
         {sliceNewsListData?.map((data: NewsItemType) => (
-          <div key={data.id}>
+          <div key={data.id} onClick={handleRead}>
             <NewsPostItem newsItem={data} />
           </div>
         ))}
